requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aioredis==2.0.1
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
basicauth==1.0.0
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
constantly==15.1.0
cryptography==41.0.3
daphne==4.0.0
defusedxml==0.7.1
dj-database-url==2.1.0
dj-rest-auth==5.0.1
Django==4.2
django-allauth==0.55.2
django-cors-headers==4.2.0
djangorestframework==3.14.0
frozenlist==1.4.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.5
multidict==6.0.4
oauthlib==3.2.2
Pillow==10.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyfcm==1.5.4
PyJWT==2.8.0
pyOpenSSL==23.2.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
stripe==6.0.0
twilio==8.9.1
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
websockets==11.0.3
whitenoise==6.6.0
yarl==1.9.2
zope.interface==6.0
